Rename Image to SanityImage for PortableText to prevent warnings
